+++ R/cpp11.R
@@ -12,6 +12,14 @@ wsAddProtocols <- function(wsc_xptr, protocols) {
   invisible(.Call(`_websocket_wsAddProtocols`, wsc_xptr, protocols))
 }
 
+wsAddProxy <- function(wsc_xptr, proxy_url) {
+  invisible(.Call(`_websocket_wsAddProxy`, wsc_xptr, proxy_url))
+}
+
+wsAddProxyBasicAuth <- function(wsc_xptr, username, password) {
+  invisible(.Call(`_websocket_wsAddProxyBasicAuth`, wsc_xptr, username, password))
+}
+
 wsConnect <- function(wsc_xptr) {
   invisible(.Call(`_websocket_wsConnect`, wsc_xptr))
 }